Remove MigrationFlowNavigationController and move navigation bar appearance configuration to MigrationFlowCoordinator
